:zap: Implement AI tree-execution view
