+++ packages/editor-ui/src/components/RunData.vue
@@ -181,11 +181,7 @@
 
 		<div
 			v-else-if="
-				hasNodeRun &&
-				dataCount > 0 &&
-				maxRunIndex === 0 &&
-				!isArtificialRecoveredEventItem &&
-				!isAiView
+				hasNodeRun && dataCount > 0 && maxRunIndex === 0 && !isArtificialRecoveredEventItem
 			"
 			v-show="!editMode.enabled"
 			:class="$style.itemsCount"
@@ -296,6 +292,9 @@
 				/>
 			</div>
 
+			<!-- V-else slot named content which only renders if $slots.content is passed and hasNodeRun -->
+			<slot name="content" v-else-if="hasNodeRun && $slots['content']"></slot>
+
 			<div
 				v-else-if="
 					hasNodeRun &&
@@ -342,10 +341,6 @@
 					:runIndex="runIndex"
 					:totalRuns="maxRunIndex"
 				/>
-			</Suspense>
-
-			<Suspense v-else-if="hasNodeRun && isPaneTypeOutput && displayMode === 'ai'">
-				<run-data-ai :node="ndvStore.activeNode" />
 			</Suspense>
 
 			<Suspense v-else-if="hasNodeRun && isPaneTypeOutput && displayMode === 'html'">
@@ -459,8 +454,7 @@
 				!hasRunError &&
 				binaryData.length === 0 &&
 				dataCount > pageSize &&
-				!isSchemaView &&
-				!isAiView
+				!isSchemaView
 			"
 			v-show="!editMode.enabled"
 		>
@@ -498,7 +492,6 @@
 import { defineAsyncComponent, defineComponent } from 'vue';
 import type { PropType } from 'vue';
 import { mapStores } from 'pinia';
-import { get } from 'lodash-es';
 import { saveAs } from 'file-saver';
 import type {
 	ConnectionTypes,
@@ -688,9 +681,6 @@ export default defineComponent({
 			}
 			return null;
 		},
-		isAiView(): boolean {
-			return this.displayMode === 'ai';
-		},
 		isSchemaView(): boolean {
 			return this.displayMode === 'schema';
 		},
@@ -737,15 +727,6 @@ export default defineComponent({
 				this.activeNode.parameters.operation === 'generateHtmlTemplate'
 			) {
 				defaults.unshift({ label: 'HTML', value: 'html' });
-			}
-
-			if (this.isPaneTypeOutput && this.activeNode) {
-				const resultData = this.workflowsStore.getWorkflowResultDataByNodeName(
-					this.activeNode.name,
-				);
-				if (get(resultData, [this.runIndex, 'metadata'])) {
-					defaults.unshift({ label: 'AI', value: 'ai' });
-				}
 			}
 
 			return defaults;
@@ -1286,18 +1267,6 @@ export default defineComponent({
 
 			return inputData.length;
 		},
-		hasAiMetadata(): boolean {
-			if (this.node) {
-				const resultData = this.workflowsStore.getWorkflowResultDataByNodeName(this.node.name);
-
-				if (!resultData || !Array.isArray(resultData)) {
-					return false;
-				}
-
-				return !!resultData[resultData.length - 1!].metadata;
-			}
-			return false;
-		},
 		init() {
 			// Reset the selected output index every time another node gets selected
 			this.outputIndex = 0;
@@ -1311,15 +1280,6 @@ export default defineComponent({
 					mode: 'binary',
 				});
 			} else if (this.displayMode === 'binary') {
-				this.ndvStore.setPanelDisplayMode({
-					pane: this.paneType as 'input' | 'output',
-					mode: 'table',
-				});
-			}
-
-			if (this.displayMode === 'ai' && !this.hasAiMetadata()) {
-				// If the user has previously selected the AI view but the
-				// current node doesn't have any data disply the table view instead
 				this.ndvStore.setPanelDisplayMode({
 					pane: this.paneType as 'input' | 'output',
 					mode: 'table',
