+++ packages/editor-ui/src/components/WorkflowLMChat.vue
@@ -60,7 +60,7 @@
 					</div>
 				</div>
 				<div class="logs">
-					<run-data-ai v-if="node" :node="node" />
+					<run-data-ai v-if="node" :node="node" hide-title slim />
 					<div v-else class="no-node-connected">
 						<n8n-text tag="div" :bold="true" color="text-dark" size="large">{{
 							$locale.baseText('chat.window.noExecution')
@@ -469,6 +469,7 @@ export default defineComponent({
 		height: 100%;
 		overflow-y: auto;
 		width: 100%;
+		padding: var(--spacing-xs) 0;
 	}
 	.messages {
 		background-color: var(--color-background-base);
