+++ packages/editor-ui/src/plugins/i18n/locales/en.json
@@ -727,6 +727,10 @@
 	"ndv.input.disabled": "The '{nodeName}' node is disabled and won’t execute.",
 	"ndv.input.disabled.cta": "Enable it",
 	"ndv.output": "Output",
+	"ndv.output.ai.empty": "👈 This is {node}’s AI Logs. Click on a node to see the input it received and data it outputted.",
+	"ndv.output.ai.logTree": "Log tree",
+	"ndv.output.outType.ai": "AI Logs",
+	"ndv.output.outType.regular": "Node Output",
 	"ndv.output.edit": "Edit Output",
 	"ndv.output.all": "all",
 	"ndv.output.branch": "Branch",
